Started on the menus.
A page with all the menus

+++ Models/Menu.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string DishName { get; set; }
         public string Description { get; set; }
-        public int DishPrice { get; set; }
+        public double DishPrice { get; set; }
     }
 }
